refactor(#8) : user -> auth로 명칭변경
