+++ spbackend/src/main/java/com/luminous/aurora/auth/repository/UserRepository.java
@@ -1,11 +1,11 @@
-package com.luminous.aurora.user.repository;
+package com.luminous.aurora.auth.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.luminous.aurora.user.entity.User;
+import com.luminous.aurora.auth.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
