+++ spbackend/src/main/java/com/luminous/aurora/jwt/JwtAuthenticationFilter.java
@@ -1,8 +1,7 @@
 package com.luminous.aurora.jwt;
 
 
-import com.luminous.aurora.user.entity.User;
-import com.luminous.aurora.user.repository.UserRepository;
+import com.luminous.aurora.auth.repository.UserRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -12,7 +11,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
